# apps/api/src/api/rules-settings/dto/rule-setting-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { ServiceCode } from '@trading-bot/models';

export class RuleSettingResponseDto {
  @ApiProperty({ example: 1 })
  id: number;

  @ApiProperty({ example: 'BTC Price Alert' })
  name: string;

  @ApiProperty({ example: 'BTC_ALERT_01' })
  code: string;

  @ApiProperty({ example: 'Sends a Telegram alert when BTC price drops below $50,000' })
  description: string;

  @ApiProperty({
    example: {
      type: 'log',
      arguments: { message: 'This message will be logged' }
    }
  })
  configuration: any;

  @ApiProperty({ example: 1 })
  authorId: number;

  @ApiProperty({ enum: ServiceCode, example: ServiceCode.TELEGRAM })
  serviceCode: ServiceCode;

  @ApiProperty({ example: ['crypto', 'binance'] })
  tags?: string[];
}