# libs/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/rules": {
      "post": {
        "operationId": "RulesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rule created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new trading rule",
        "tags": [
          "rules"
        ]
      },
      "get": {
        "operationId": "RulesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rules retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRulesDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all rules for the authenticated user with pagination",
        "tags": [
          "rules"
        ]
      }
    },
    "/api/v1/rules/{id}": {
      "get": {
        "operationId": "RulesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Rule ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rule retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Rule not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific rule by ID",
        "tags": [
          "rules"
        ]
      },
      "patch": {
        "operationId": "RulesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Rule ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rule updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Rule not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a rule",
        "tags": [
          "rules"
        ]
      },
      "delete": {
        "operationId": "RulesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Rule ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rule deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Rule not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a rule",
        "tags": [
          "rules"
        ]
      }
    },
    "/api/v1/rules-settings": {
      "post": {
        "operationId": "RulesSettingsController_createSetting",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRuleSettingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Setting created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSettingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new universal rule setting",
        "tags": [
          "rules-settings"
        ]
      },
      "get": {
        "operationId": "RulesSettingsController_findAllSettings",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "serviceCode",
            "required": false,
            "in": "query",
            "description": "Filter by service code",
            "schema": {
              "enum": [
                "BINANCE",
                "BYBIT",
                "KRAKEN",
                "TELEGRAM",
                "EMAIL",
                "DISCORD_WEBHOOKS",
                "SLACK_WEBHOOKS",
                "SMS_TWILIO",
                "PUSH_NOTIFICATIONS_ONESIGNAL",
                "WHATSAPP_BUSINESS",
                "WEBHOOKS"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of rule settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleSettingResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all universal rule settings for user",
        "tags": [
          "rules-settings"
        ]
      }
    },
    "/api/v1/rules-settings/{id}": {
      "patch": {
        "operationId": "RulesSettingsController_updateSetting",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRuleSettingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Setting updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSettingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a rule setting",
        "tags": [
          "rules-settings"
        ]
      },
      "delete": {
        "operationId": "RulesSettingsController_removeSetting",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Setting deleted successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a rule setting",
        "tags": [
          "rules-settings"
        ]
      }
    },
    "/api/v1/rules-settings/telegram-chat-id/{id}": {
      "get": {
        "operationId": "RulesSettingsController_getTelegramChatId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Rule setting ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Telegram chat ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramChatIdResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Telegram chat ID for a rule setting",
        "tags": [
          "rules-settings"
        ]
      }
    },
...
  },
  "info": {
    "title": "Trading Bot API",
    "description": "API for trading bot user management and authentication",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RuleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Unique identifier of the rule",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Name of the trading rule",
            "example": "BTC Price Alert"
          },
          "description": {
            "type": "string",
            "description": "Description of what this rule does",
            "example": "Sends a Telegram alert when BTC price drops below $50,000"
          },
          "ruleBody": {
            "type": "object",
            "description": "Rule configuration as JSON object",
            "example": {
              "type": "log",
              "arguments": {
                "message": "This message will be logged"
              }
            }
          },
          "authorId": {
            "type": "number",
            "description": "ID of the user who created this rule",
            "example": 1
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "ruleBody",
          "authorId"
        ]
      },
      "PaginatedRulesDto": {
        "type": "object",
        "properties": {
          "rules": {
            "description": "Array of rules for the requested page",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 120,
            "description": "Total number of rules available"
          }
        },
        "required": [
          "rules",
          "total"
        ]
      },
      "CreateRuleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the trading rule",
            "example": "BTC Price Alert"
          },
          "description": {
            "type": "string",
            "description": "Description of what this rule does",
            "example": "Sends a Telegram alert when BTC price drops below $50,000"
          },
          "ruleBody": {
            "type": "object",
            "description": "Rule configuration as JSON object",
            "example": {
              "type": "log",
              "arguments": {
                "message": "This message will be logged"
              }
            }
          }
        },
        "required": [
          "name",
          "description",
          "ruleBody"
        ]
      },
      "UpdateRuleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the trading rule",
            "example": "BTC Price Alert"
          },
          "description": {
            "type": "string",
            "description": "Description of what this rule does",
            "example": "Sends a Telegram alert when BTC price drops below $50,000"
          },
          "ruleBody": {
            "type": "object",
            "description": "Rule configuration as JSON object",
            "example": {
              "type": "log",
              "arguments": {
                "message": "This message will be logged"
              }
            }
          }
        }
      },
      "CreateUserRuleSettingDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "My Binance Bot"
          },
          "code": {
            "type": "string",
            "example": "BINANCE_MAIN_01"
          },
          "description": {
            "type": "string",
            "example": "Rule for spot trading"
          },
          "serviceCode": {
            "type": "string",
            "enum": [
              "BINANCE",
              "BYBIT",
              "KRAKEN",
              "TELEGRAM",
              "EMAIL",
              "DISCORD_WEBHOOKS",
              "SLACK_WEBHOOKS",
              "SMS_TWILIO",
              "PUSH_NOTIFICATIONS_ONESIGNAL",
              "WHATSAPP_BUSINESS",
              "WEBHOOKS"
            ],
            "example": "TELEGRAM"
          },
          "tags": {
            "example": [
              "crypto",
              "binance"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "configuration": {
            "type": "object",
            "example": {
              "ApiKey": "...",
              "ApiSecret": "...",
              "BaseUrl": "..."
            }
          }
        },
        "required": [
          "name",
          "code",
          "serviceCode",
          "configuration"
        ]
      },
      "RuleSettingResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "BTC Price Alert"
          },
          "code": {
            "type": "string",
            "example": "BTC_ALERT_01"
          },
          "description": {
            "type": "string",
            "example": "Sends a Telegram alert when BTC price drops below $50,000"
          },
          "configuration": {
            "type": "object",
            "example": {
              "type": "log",
              "arguments": {
                "message": "This message will be logged"
              }
            }
          },
          "authorId": {
            "type": "number",
            "example": 1
          },
          "serviceCode": {
            "type": "string",
            "enum": [
              "BINANCE",
              "BYBIT",
              "KRAKEN",
              "TELEGRAM",
              "EMAIL",
              "DISCORD_WEBHOOKS",
              "SLACK_WEBHOOKS",
              "SMS_TWILIO",
              "PUSH_NOTIFICATIONS_ONESIGNAL",
              "WHATSAPP_BUSINESS",
              "WEBHOOKS"
            ],
            "example": "TELEGRAM"
          },
          "tags": {
            "example": [
              "crypto",
              "binance"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "name",
          "code",
          "description",
          "configuration",
          "authorId",
          "serviceCode",
          "tags"
        ]
      },
      "UpdateUserRuleSettingDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "My Binance Bot"
          },
          "code": {
            "type": "string",
            "example": "BINANCE_MAIN_01"
          },
          "description": {
            "type": "string",
            "example": "Rule for spot trading"
          },
          "serviceCode": {
            "type": "string",
            "enum": [
              "BINANCE",
              "BYBIT",
              "KRAKEN",
              "TELEGRAM",
              "EMAIL",
              "DISCORD_WEBHOOKS",
              "SLACK_WEBHOOKS",
              "SMS_TWILIO",
              "PUSH_NOTIFICATIONS_ONESIGNAL",
              "WHATSAPP_BUSINESS",
              "WEBHOOKS"
            ],
            "example": "TELEGRAM"
          },
          "tags": {
            "example": [
              "crypto",
              "binance"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "configuration": {
            "type": "object",
            "example": {
              "ApiKey": "...",
              "ApiSecret": "...",
              "BaseUrl": "..."
            }
          }
        }
      },
      "TelegramChatIdResponseDto": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "number",
            "example": 1234567890
          }
        },
        "required": [
          "chatId"
        ]
      },
...
    }
  }
}